+++ frontend/components/commons/user/UserAvatarTooltip.vue
@@ -19,15 +19,19 @@
       </div>
       <div class="">
         <NuxtLink class="user__link" :to="{ name: 'user-settings' }">
-          My settings
+          {{ $t("userAvatarTooltip.settings") }}
         </NuxtLink>
         <a
           class="user__link"
           :href="$config.documentationSite"
           target="_blank"
-          v-text="'View docs'"
+          v-text="$t('userAvatarTooltip.docs')"
         />
-        <a class="user__link" @click.prevent="logout" v-text="'Log out'" />
+        <a
+          class="user__link"
+          @click.prevent="logout"
+          v-text="$t('userAvatarTooltip.logout')"
+        />
       </div>
       <span class="copyright"
         >© {{ currentYear }} Argilla ({{ $config.clientVersion }})</span