+++ frontend/components/feedback-task/container/fields/span-annotation/components/EntityComponent.vue
@@ -1,52 +1,81 @@
 <template>
-  <div
-    class="span-entity__wrapper"
-    :style="{ left: entityPosition.left, top: entityPosition.top }"
-    ref="spanEntityRef"
-    id="spanEntity"
-  >
-    <div
-      v-on="!singleOption ? { click: toggleDropdown } : {}"
-      :class="!singleOption ? 'span-entity--clickable' : 'span-entity'"
-      v-if="!visibleDropdown"
-    >
-      <BaseButton
-        class="span-entity__close-button"
-        @click="removeSelectedOption"
-      >
-        <svgicon
-          class="span-entity__close-button__icon"
-          name="close"
-          width="10"
-          height="10"
-      /></BaseButton>
-      <span class="span-entity__text" v-text="selectedOption.text" />
-      <svgicon
-        v-if="!!suggestion"
-        class="span-entity__suggestion"
-        name="suggestion"
-        width="8"
-        height="8"
-      />
+  <span class="span-entity__wrapper">
+    <template v-if="allowOverlapping">
       <span
-        v-if="suggestionScore"
-        class="span-entity__score"
-        v-text="suggestionScore"
-      />
+        v-for="(line, index) in lines"
+        :key="index"
+        class="span-entity__line"
+        :style="{
+          width: line.width,
+          left: line.left,
+          top: line.top,
+        }"
+      ></span>
+    </template>
+    <div
+      class="span-entity__container"
+      :style="entityRelativePosition"
+      ref="spanEntityRef"
+      id="spanEntity"
+    >
+      <div
+        v-on="!singleOption ? { click: toggleDropdown } : {}"
+        @mouseenter="hoverSpan(true)"
+        @mouseleave="hoverSpan(false)"
+        :class="[
+          !singleOption ? 'span-entity--clickable' : 'span-entity',
+          allowOverlapping ? 'span-entity--overlapping' : null,
+        ]"
+        v-if="!visibleDropdown"
+      >
+        <BaseButton
+          class="span-entity__close-button"
+          @click.stop="removeSelectedSpan(span)"
+        >
+          <svgicon name="close" width="10" height="10" color="#fff"
+        /></BaseButton>
+        <span class="span-entity__text"
+          ><span v-text="selectedOption.text"
+        /></span>
+        <svgicon
+          v-if="!!suggestion"
+          :class="
+            suggestionScore
+              ? 'span-entity__suggestion--score'
+              : 'span-entity__suggestion'
+          "
+          name="suggestion"
+          width="8"
+          height="8"
+        />
+        <span
+          v-if="suggestionScore"
+          class="span-entity__score"
+          v-text="suggestionScore"
+        />
+      </div>
+      <template v-else>
+        <EntityDropdownOverlapping
+          v-if="allowOverlapping"
+          :style="overlappingDropdownPosition"
+          :selectedOption="selectedOption"
+          :options="options"
+          :spanInRange="spanInRange"
+          @on-add-span-base-on="addSpanBaseOn"
+          @on-remove-span="removeSelectedSpan"
+          v-click-outside="hideDropdown"
+        />
+        <EntityDropdown
+          v-else
+          :style="entityFixedPosition"
+          :selectedOption="selectedOption"
+          :options="options"
+          @on-replace-entity="selectEntity"
+          v-click-outside="hideDropdown"
+        />
+      </template>
     </div>
-    <EntityComponentDropdown
-      v-else
-      :style="{
-        left: spanEntityPosition.left,
-        top: spanEntityPosition.top,
-      }"
-      :selectedOption="selectedOption"
-      :options="options"
-      @on-replace-option="selectOption"
-      @on-remove-option="removeSelectedOption"
-      v-click-outside="hideDropdown"
-    />
-  </div>
+  </span>
 </template>
 
 <script>
@@ -56,12 +85,17 @@ import "assets/icons/suggestion";
 export default {
   name: "EntityComponent",
   props: {
+    span: {
+      type: Object,
+      required: true,
+    },
     entity: {
       type: Object,
       required: true,
     },
     spanQuestion: {
       type: Object,
+      required: true,
     },
     suggestion: {
       type: Object,
@@ -70,13 +104,17 @@ export default {
       type: Object,
       required: true,
     },
+    spanInRange: {
+      type: Array,
+      required: true,
+    },
   },
   data() {
     return {
       visibleDropdown: false,
       spanEntityPosition: {
-        left: "0px",
-        top: "0px",
+        left: 0,
+        top: 0,
       },
     };
   },
@@ -93,15 +131,85 @@ export default {
     suggestionScore() {
       return this.suggestion?.score?.toFixed(1);
     },
+    entityColor() {
+      return this.entity.color;
+    },
+    allowOverlapping() {
+      return this.spanQuestion.settings.allow_overlapping;
+    },
+    lines() {
+      const lines = [];
+      const { width, top, topEnd, left, right } = this.entityPosition;
+      const space = topEnd - top;
+      const linesCount = space > 0 ? this.getNumberOfLines(space) : 0;
+
+      for (let i = 0; i <= linesCount; i++) {
+        lines.push({
+          width: linesCount
+            ? i === 0
+              ? `${width - left}px`
+              : i === linesCount
+              ? `${right}px`
+              : `${width}px`
+            : `${width}px`,
+          left: i === 0 ? `${left}px` : 0,
+          top: `${top + i * this.entityPosition.lineHeight}px`,
+        });
+      }
+
+      return lines;
+    },
+    entityWidth() {
+      return `${this.entityPosition.width}px`;
+    },
+    entityRelativePosition() {
+      return {
+        left: `${this.entityPosition.left}px`,
+        top: `${this.entityPosition.top}px`,
+      };
+    },
+    entityFixedPosition() {
+      return {
+        left: `${this.spanEntityPosition.left}px`,
+        top: `${this.spanEntityPosition.top}px`,
+      };
+    },
+    getLevelInEntitiesInRange() {
+      return this.spanInRange.findIndex(
+        ({ entity }) => entity.id === this.entity.id
+      );
+    },
+    overlappingDropdownPosition() {
+      return {
+        left: `${this.spanEntityPosition.left}px`,
+        top: `${
+          this.spanEntityPosition.top -
+          this.getLevelInEntitiesInRange * this.entityPosition.baseEntityGap
+        }px`,
+      };
+    },
   },
   methods: {
-    selectOption(option) {
-      this.$emit("on-replace-option", option);
+    getNumberOfLines(space) {
+      return Math.floor(space / this.entityPosition.lineHeight + 1);
+    },
+    addSpanBaseOn(span, entity) {
+      this.$emit("on-add-span-base-on", span, entity);
       this.hideDropdown();
     },
-    removeSelectedOption() {
-      this.$emit("on-remove-option");
+    selectEntity(entity) {
+      this.$emit("on-replace-entity", entity);
       this.hideDropdown();
+    },
+    removeSelectedSpan(span) {
+      this.$emit("on-remove-span", span);
+      this.hideDropdown();
+    },
+    hoverSpan(isHovered) {
+      this.getPosition();
+      if (this.allowOverlapping) {
+        this.$emit("on-hover-span", isHovered);
+      }
     },
     toggleDropdown() {
       this.visibleDropdown = !this.visibleDropdown;
@@ -112,18 +220,17 @@ export default {
     },
     getPosition() {
       const position = this.$refs.spanEntityRef.getBoundingClientRect();
-      this.spanEntityPosition.left = `${position.left}px`;
-      this.spanEntityPosition.top = `${
-        position.top + this.$refs.spanEntityRef.scrollTop
-      }px`;
+      this.spanEntityPosition.left = position.left;
+      this.spanEntityPosition.top =
+        position.top + this.$refs.spanEntityRef.scrollTop;
     },
     getScrollParent(element) {
       if (!element) return;
 
       let parent = element.parentElement;
       while (parent) {
-        const { overflow } = window.getComputedStyle(parent);
-        if (overflow.split(" ").every((o) => o === "auto" || o === "scroll")) {
+        const { "overflow-y": overflowY } = window.getComputedStyle(parent);
+        if (overflowY.split(" ").every((o) => o === "auto" || o === "scroll")) {
           return parent;
         }
 
@@ -134,11 +241,12 @@ export default {
     },
   },
   mounted() {
-    this.scroll = this.getScrollParent(document.getElementById("spanEntity"));
-
-    if (this.scroll) {
-      this.scroll.addEventListener("scroll", this.getPosition);
-    }
+    this.$nextTick(() => {
+      this.scroll = this.getScrollParent(document.getElementById("spanEntity"));
+      if (this.scroll) {
+        this.scroll.addEventListener("scroll", this.getPosition);
+      }
+    });
   },
   beforeDestroy() {
     if (this.scroll) {
@@ -153,15 +261,13 @@ export default {
 .span-entity {
   $this: &;
   display: flex;
-  gap: 2px;
   align-items: center;
   flex-shrink: 0;
-  min-width: 10px;
-  max-width: v-bind("entityPosition.width");
+  min-width: $base-space * 2;
+  max-width: v-bind(entityWidth);
   margin-top: -1px;
-  padding-right: 3px;
   text-transform: uppercase;
-  font-family: "Roboto Condensed", sans-serif;
+  font-family: $tertiary-font-family;
   user-select: none;
   transform-origin: left top;
   transition: scale 0.2s;
@@ -176,24 +282,47 @@ export default {
   }
   &__wrapper {
     position: absolute;
+    pointer-events: none;
+  }
+  &__container {
+    position: absolute;
     display: flex;
     margin-top: 20px;
     line-height: 1.2;
     @include font-size(11px);
+    pointer-events: all;
   }
   &__score {
-    display: none;
-    margin-left: 2px;
+    display: inline-flex;
+    flex-shrink: 0;
+    margin-right: calc($base-space / 2);
   }
   &__text {
-    gap: 4px;
+    margin-inline: calc($base-space / 4);
     @include truncate(auto);
+    @include font-size(0px);
+    span {
+      @include font-size(11px);
+    }
+  }
+  &__line {
+    position: absolute;
+    height: 2px;
+    background: v-bind(entityColor);
+    margin-top: 20px;
   }
   &__suggestion {
-    flex-shrink: 0;
+    margin-right: calc($base-space / 2);
+    &--score {
+      display: none;
+      margin-right: calc($base-space / 2);
+      flex-shrink: 0;
+    }
   }
   &:hover {
-    position: relative;
+    position: fixed;
+    left: v-bind("entityFixedPosition.left");
+    top: v-bind("entityFixedPosition.top");
     z-index: 1;
     transition: scale 0.2s;
     max-width: none;
@@ -203,9 +332,15 @@ export default {
     }
     #{$this}__text {
       min-width: $base-space * 3;
+      max-width: none !important;
     }
     #{$this}__score {
       display: inline-flex;
+    }
+    #{$this}__suggestion {
+      &--score {
+        display: inline-flex;
+      }
     }
     &:before {
       background: v-bind("selectedOption.color");
@@ -215,23 +350,21 @@ export default {
   }
   &__close-button {
     display: none;
-    height: 100%;
-    padding: 0 1px;
-    margin-right: calc($base-space / 2);
-    flex-shrink: 0;
+    padding: 1px;
     border-radius: 0;
     background: $black-54;
     &:hover {
       background: $black-87;
     }
-    &__icon {
-      min-width: 10px;
-      color: palette(white);
-    }
   }
   &--clickable {
     cursor: pointer;
     @extend .span-entity;
+  }
+  &--overlapping {
+    @extend .span-entity;
+    background: v-bind(entityColor);
+    margin-top: 0;
   }
 }
 </style>