+++ frontend/nuxt.config.ts
@@ -124,11 +124,14 @@ const config: NuxtConfig = {
         code: "en",
         file: "en.js",
       },
+      {
+        code: "de",
+        file: "de.js",
+      },
     ],
-    detectBrowserLanguage: {
-      useCookie: true,
-      cookieKey: "i18n_redirected",
-      redirectOn: "root",
+    detectBrowserLanguage: false,
+    vueI18n: {
+      fallbackLocale: "en",
     },
     lazy: true,
     langDir: "translation/",
@@ -202,6 +205,7 @@ const config: NuxtConfig = {
         },
       },
     },
+    cookie: false,
     resetOnError: true,
     redirect: { login: "/sign-in", logout: "/sign-in" },
   },