+++ frontend/v1/domain/usecases/dataset-setting/update-dataset-setting-use-case.ts
@@ -1,8 +1,8 @@
-import { Dataset } from "../../entities/Dataset";
-import { DatasetRepository } from "~/v1/infrastructure/repositories";
+import { Dataset } from "../../entities/dataset/Dataset";
+import { IDatasetRepository } from "../../services/IDatasetRepository";
 
 export class UpdateDatasetSettingUseCase {
-  constructor(private readonly datasetRepository: DatasetRepository) {}
+  constructor(private readonly datasetRepository: IDatasetRepository) {}
 
   async execute(dataset: Dataset) {
     const response = await this.datasetRepository.update(dataset);